+++ app/publy-backend/src/main/java/nh/publy/backend/graphql/StoryConnection.java
@@ -5,8 +5,8 @@ import nh.publy.backend.graphql.sql.SelectCountSqlBuilder;
 import nh.publy.backend.graphql.sql.SelectNodesSqlBuilder;
 import nh.publy.backend.graphql.sql.SqlOrder;
 
-import javax.persistence.EntityManager;
-import javax.persistence.Query;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.Query;
 import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
@@ -67,8 +67,8 @@ public class StoryConnection extends AbstractPublyConnection<Story> {
   }
 
   @Override
-  protected BigInteger loadTotalCount() {
+  protected Long loadTotalCount() {
     Query countQuery = new SelectCountSqlBuilder("stories", this.condition).buildCountQuery(entityManager);
-    return (BigInteger) countQuery.getSingleResult();
+    return (Long) countQuery.getSingleResult();
   }
 }
